Fri, Apr 26, 2024, 12:43 PM +08:00

+++ L4/ps4/filter-less/helpers.c
@@ -9,7 +9,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-            float average = (image[i][j].rgbtBlue + image[i][j].rgbtGreen + image[i][j].rgbtRed) / 3;
+            float average = (float)(image[i][j].rgbtBlue + image[i][j].rgbtGreen + image[i][j].rgbtRed) / 3;
 
             image[i][j].rgbtBlue = (int)round(average);
             image[i][j].rgbtGreen = (int)round(average);
